+++ marker_api/src/ast/item.rs
@@ -10,8 +10,8 @@ mod mod_item;
 pub use mod_item::ModItem;
 mod static_item;
 pub use self::static_item::StaticItem;
-mod use_decl_item;
-pub use self::use_decl_item::*;
+mod use_item;
+pub use self::use_item::*;
 mod const_item;
 pub use self::const_item::ConstItem;
 mod fn_item;
